chore: refactor out utils and middlware
[ci skip]
