+++ src/middleware.ts
@@ -4,12 +4,12 @@ import type {
   JSONRPCResponse,
   JSONRPCResponseResult,
   MiddlewareFactory,
-} from '../types';
+} from './types';
 import { TransformStream } from 'stream/web';
 import { JSONParser } from '@streamparser/json';
 import * as rpcUtils from './utils';
 import { promise } from './utils';
-import * as rpcErrors from '../errors';
+import * as rpcErrors from './errors';
 
 /**
  * This function is a factory to create a TransformStream that will
