+++ src/utils/index.ts
@@ -1,2 +1,2 @@
 export * from './middleware';
-export * from './utils';
+export * from '../utils';
