+++ tests/utils/middleware.test.ts
@@ -6,7 +6,7 @@ import { AsyncIterableX as AsyncIterable } from 'ix/asynciterable';
 import * as rpcUtils from '@/utils';
 import 'ix/add/asynciterable-operators/toarray';
 import * as rpcErrors from '@/errors';
-import * as rpcUtilsMiddleware from '@/utils/middleware';
+import * as rpcUtilsMiddleware from '@/middleware';
 import * as rpcTestUtils from '../utils';
 
 describe('Middleware tests', () => {
